Header.html - Page Redirection Added
I have created "projectSection" division in the homepage and "contactAboutPage" in the about page. The header redirection is added. However, the alignment of the project and contact sections are off upon adding the divisions/sections from the .jsp files.

I do not know how the styling work, so could not identify the causing of change in the alignment.